Fixed major bug with tab focus
The next item that needs to be fixed is the additional tabs generated from certain Easy Apply's like recruiters.  This should be an additional script that fires to take care of that application; however, this is more of a data science issue and will need to be fault tolerant in order to harvest most of the questions from the "sub form" or sub application generated by Linkedin.